fix(frontend/controller): fix search highlighting transliterations
Refactors Results into single result component
